#32 code improvements
