+++ wirecardpaymentgateway/models/Wirecard/PaymentGateway/Payment.php
@@ -31,20 +31,13 @@
 
 require_once __DIR__.'/../../../vendor/autoload.php';
 require_once __DIR__.'/../../../libraries/Logger.php';
+require_once __DIR__.'/Cart.php';
 
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
 use Wirecard\PaymentSdk\Response\FailureResponse;
 use Wirecard\PaymentSdk\Response\InteractionResponse;
 use Wirecard\PaymentSdk\TransactionService;
-use Wirecard\PaymentSdk\Entity\AccountHolder;
-use Wirecard\PaymentSdk\Entity\Basket;
-use Wirecard\PaymentSdk\Entity\Item;
-use Wirecard\PaymentSdk\Entity\CustomField;
-use Wirecard\PaymentSdk\Entity\CustomFieldCollection;
-use Wirecard\PaymentSdk\Entity\Redirect;
-use Wirecard\PaymentSdk\Entity\Device;
-use Wirecard\PaymentSdk\Entity\Amount;
 
 class WirecardPaymentGatewayPayment
 {
@@ -132,24 +125,26 @@ class WirecardPaymentGatewayPayment
 
     public function initiate($cart, $orderNumber)
     {
+        $CartData=new WirecardPaymentGatewayCart($this->module);
         $transaction =$this->getTransaction($cart, $orderNumber);
-        $transaction->setRedirect($this->getRedirect($cart, $orderNumber));
-        $transaction->setNotificationUrl($this->getNotification($cart, $orderNumber));
-        $transaction->setAmount($this->getTotalAmount($cart));
+        $transaction->setRedirect($CartData->getRedirect($cart, $orderNumber));
+        $transaction->setNotificationUrl($CartData->getNotification($cart, $orderNumber));
+        $transaction->setAmount($CartData->getTotalAmount($cart));
         $transaction->setDescriptor($this->getDescriptor($orderNumber));
 
         $transaction->setConsumerId($cart->id_customer);
-        $transaction->setIpAddress($this->getConsumerIpAddress());
-        $transaction->setAccountHolder($this->getConsumerData($cart));
-        $transaction->setShipping($this->getShippingData($cart));
-        $transaction->setDevice($this->getDevice($cart->id_customer));
-        $transaction->setCustomFields($this->setCustomField(array('customOrderNumber'=>$orderNumber)));
+        $transaction->setIpAddress($CartData->getConsumerIpAddress());
+        $transaction->setAccountHolder($CartData->getConsumerData($cart));
+        $transaction->setShipping($CartData->getShippingData($cart));
+        $transaction->setDevice($CartData->getDevice($cart->id_customer));
+        $transaction->setCustomFields($CartData->setCustomField(array('customOrderNumber'=>$orderNumber)));
 
         $logger = new Logger();
         $transactionService = new TransactionService($this->getConnection(), $logger);
         $response = $transactionService->pay($transaction);
         $this->processResponse($response);
     }
+
 
     public function processResponse($response)
     {
@@ -175,184 +170,6 @@ class WirecardPaymentGatewayPayment
                 throw new ExceptionEE($message);
             }
         }
-    }
-
-    public function getUrlParameters($cart, $orderNumber)
-    {
-        return array(
-            'id_cart' => (int)$cart->id,
-            'id_module' => (int)$this->module->id,
-            'key' => $cart->secure_key,
-            'order' => $orderNumber
-        );
-    }
-
-    public function getRedirect($cart, $orderNumber)
-    {
-        $params=$this->getUrlParameters($cart, $orderNumber);
-        return new Redirect(
-            $this->module->getContext()->link->getModuleLink($this->module->getName(), 'success', $params, true),
-            $this->module->getContext()->link->getModuleLink($this->module->getName(), 'cancel', $params, true)
-        );
-    }
-
-    public function getNotification($cart, $orderNumber)
-    {
-        $params=$this->getUrlParameters($cart, $orderNumber);
-        return $this->module->getContext()->link->getModuleLink(
-            $this->module->getName(),
-            'notify',
-            $params,
-            true
-        );
-    }
-
-    protected function getConsumerData($cart)
-    {
-        $customer = new Customer($cart->id_customer);
-
-        $customerData = new AccountHolder();
-        $customerData->setFirstName($customer->firstname);
-        $customerData->setLastName($customer->lastname);
-        $customerData->setEmail($customer->email);
-        $customerData->setGender($customer->id_gender);
-        if (Tools::strlen($customer->birthday) && $customer->birthday !== "0000-00-00") {
-            $birthday = new DateTime($customer->birthday);
-            $customerData->setDateOfBirth($birthday);
-        }
-
-        return $customerData;
-    }
-
-    protected function getShippingData($cart)
-    {
-        $carrier = new Carrier($cart->id_carrier);
-        $addressDelivery = new Address(intval($cart->id_address_delivery));
-
-        $shippingData = new AccountHolder();
-        $shippingData->setFirstName($addressDelivery->firstname);
-        $shippingData->setLastName($addressDelivery->lastname);
-        $shippingData->setAddress($this->getAddress($addressDelivery));
-        $shippingData->setShippingMethod($carrier->getShippingMethod());
-
-        return $shippingData;
-    }
-
-    public function getDevice($id_customer)
-    {
-        $Device=new Device();
-        $Device->setFingerprint(md5($id_customer . "_" . microtime()));
-        return $Device;
-    }
-
-    public function getBasket($cart)
-    {
-        $currency = new CurrencyCore($cart->id_currency);
-        $currencyIsoCode = $currency->iso_code;
-        $basket = new Basket();
-
-        foreach ($cart->getProducts() as $product) {
-            $price_wt=$product['price_wt'];
-            $price=$product['price'];
-            $tax = ($price_wt - $price) * 100 / $price;
-
-            $productInfo = new Item(
-                $product['name'],
-                new Amount(
-                    number_format(
-                        $price_wt,
-                        2,
-                        '.',
-                        ''
-                    ),
-                    $currencyIsoCode
-                ),
-                $product['cart_quantity']
-            );
-            $productInfo->setDescription(
-                Tools::substr(
-                    strip_tags($product['description_short']),
-                    0,
-                    127
-                )
-            );
-            $productInfo->setTaxRate(
-                number_format(
-                    $tax,
-                    2,
-                    '.',
-                    ''
-                )
-            );
-            $basket->add($productInfo);
-        }
-
-        if ($cart->getTotalShippingCost() != 0) {
-            $shipping = new Item(
-                'Shipping',
-                new Amount(
-                    number_format(
-                        $cart->getTotalShippingCost(),
-                        2,
-                        '.',
-                        ''
-                    ),
-                    $currencyIsoCode
-                ),
-                '1'
-            );
-
-            $shipping->setDescription($this->module->l('Shipping'));
-            $shipping->setTaxRate(
-                number_format(
-                    '0',
-                    2,
-                    '.',
-                    ''
-                )
-            );
-            $basket->add($shipping);
-        }
-        return $basket;
-    }
-
-    protected function getConsumerIpAddress()
-    {
-        return Tools::getRemoteAddr();
-    }
-
-    public function getTotalAmount($cart)
-    {
-        $currency = new CurrencyCore($cart->id_currency);
-        $currencyIsoCode = $currency->iso_code;
-        return new Amount($cart->getOrderTotal(true), $currencyIsoCode);
-    }
-
-    public function setCustomField($CustomFieldArray)
-    {
-        $customFields = new CustomFieldCollection();
-        if (!empty($CustomFieldArray)) {
-            foreach ($CustomFieldArray as $key => $field) {
-                $customOrderNumber = new CustomField($key, $field);
-                $customFields->add($customOrderNumber);
-            }
-        }
-        return $customFields;
-    }
-
-    protected function getAddress($source)
-    {
-        $country = new Country($source->id_country);
-
-        $address = new \Wirecard\PaymentSdk\Entity\Address(
-            $country->iso_code,
-            $source->city,
-            $source->address1
-        );
-        $address->setPostalCode($source->postcode);
-        $address->setStreet2($source->address2);
-
-        return $address;
     }
 
     public function getTransaction($cart, $orderNumber)
